normalize namespaces / drop backwards compatibility to old release mechanism (to ensure clean fresh-start)

+++ src/main/java/org/humanbrainproject/knowledgegraph/commons/nexusToArangoIndexing/control/NexusArangoTransaction.java
@@ -5,6 +5,7 @@ import com.arangodb.entity.CollectionType;
 import org.humanbrainproject.knowledgegraph.commons.jsonld.control.JsonTransformer;
 import org.humanbrainproject.knowledgegraph.commons.nexus.control.NexusDocumentConverter;
 import org.humanbrainproject.knowledgegraph.commons.nexus.control.SystemNexusClient;
+import org.humanbrainproject.knowledgegraph.commons.propertyGraph.arango.control.ArangoConnection;
 import org.humanbrainproject.knowledgegraph.commons.propertyGraph.arango.control.ArangoDocumentConverter;
 import org.humanbrainproject.knowledgegraph.commons.propertyGraph.arango.control.ArangoRepository;
 import org.humanbrainproject.knowledgegraph.commons.propertyGraph.arango.entity.ArangoDocumentReference;
@@ -27,7 +28,7 @@ import java.util.List;
 
 @Component
 @Primary
-public class NexusArangoTransaction implements DatabaseTransaction<ArangoDatabase> {
+public class NexusArangoTransaction implements DatabaseTransaction {
 
     @Autowired
     ArangoRepository repository;
@@ -50,29 +51,35 @@ public class NexusArangoTransaction implements DatabaseTransaction<ArangoDatabas
     protected Logger logger = LoggerFactory.getLogger(NexusArangoTransaction.class);
 
     @Override
-    public void execute(TodoList<ArangoDatabase> todoList) {
+    public void execute(TodoList todoList) {
         //First remove instances
-        List<DeleteTodoItem<ArangoDatabase>> deleteItems = todoList.getDeleteTodoItems();
-        for (DeleteTodoItem<ArangoDatabase> deleteItem : deleteItems) {
-            ArangoDatabase database = deleteItem.getDatabaseConnection().getOrCreateDB();
-            ArangoDocumentReference reference = ArangoDocumentReference.fromVertexOrEdgeReference(deleteItem.getReference());
-            repository.deleteDocument(reference, database);
+        List<DeleteTodoItem> deleteItems = todoList.getDeleteTodoItems();
+        for (DeleteTodoItem deleteItem : deleteItems) {
+            ArangoConnection databaseConnection = deleteItem.getDatabaseConnection(ArangoConnection.class);
+            if(databaseConnection!=null) {
+                ArangoDatabase database = databaseConnection.getOrCreateDB();
+                ArangoDocumentReference reference = ArangoDocumentReference.fromVertexOrEdgeReference(deleteItem.getReference());
+                repository.deleteDocument(reference, database);
+            }
         }
 
         //then add new instances
-        List<InsertTodoItem<ArangoDatabase>> insertItems = todoList.getInsertTodoItems();
-        for (InsertTodoItem<ArangoDatabase> insertItem : insertItems) {
-            ArangoDatabase database = insertItem.getDatabaseConnection().getOrCreateDB();
-            ArangoDocumentReference reference = ArangoDocumentReference.fromVertexOrEdgeReference(insertItem.getObject());
-            String jsonFromVertexOrEdge = arangoDocumentConverter.createJsonFromVertexOrEdge(reference, insertItem.getObject());
-            if (jsonFromVertexOrEdge != null) {
-                repository.insertDocument(reference, jsonFromVertexOrEdge, insertItem.getObject() instanceof Edge ? CollectionType.EDGES : CollectionType.DOCUMENT, database);
+        List<InsertTodoItem> insertItems = todoList.getInsertTodoItems();
+        for (InsertTodoItem insertItem : insertItems) {
+            ArangoConnection databaseConnection = insertItem.getDatabaseConnection(ArangoConnection.class);
+            if(databaseConnection!=null) {
+                ArangoDatabase database = databaseConnection.getOrCreateDB();
+                ArangoDocumentReference reference = ArangoDocumentReference.fromVertexOrEdgeReference(insertItem.getObject());
+                String jsonFromVertexOrEdge = arangoDocumentConverter.createJsonFromVertexOrEdge(reference, insertItem.getObject());
+                if (jsonFromVertexOrEdge != null) {
+                    repository.insertDocument(reference, jsonFromVertexOrEdge, insertItem.getObject() instanceof Edge ? CollectionType.EDGES : CollectionType.DOCUMENT, database);
+                }
             }
         }
 
         //and finally trigger primary store insertions/updates.
-        List<InsertOrUpdateInPrimaryStoreTodoItem<ArangoDatabase>> insertOrUpdateInPrimaryStoreItems = todoList.getInsertOrUpdateInPrimaryStoreTodoItems();
-        for (InsertOrUpdateInPrimaryStoreTodoItem<ArangoDatabase> insertOrUpdateInPrimaryStoreItem : insertOrUpdateInPrimaryStoreItems) {
+        List<InsertOrUpdateInPrimaryStoreTodoItem> insertOrUpdateInPrimaryStoreItems = todoList.getInsertOrUpdateInPrimaryStoreTodoItems();
+        for (InsertOrUpdateInPrimaryStoreTodoItem insertOrUpdateInPrimaryStoreItem : insertOrUpdateInPrimaryStoreItems) {
             MainVertex mainVertex = insertOrUpdateInPrimaryStoreItem.getObject();
             String jsonFromVertexOrEdge = nexusDocumentConverter.createJsonFromVertex(mainVertex.getInstanceReference(), mainVertex);
             NexusInstanceReference newReference = nexusClient.createOrUpdateInstance(mainVertex.getInstanceReference().setRevision(null), jsonTransformer.parseToMap(jsonFromVertexOrEdge));
